+++ CHANGES.md
@@ -30,6 +30,9 @@ Unreleased
 - Prevent crash if absolute paths are used in the install stanza and in
   recursive globs. These cases now result in a user error. (#6331, @gridbugs)
 
+- Add `(glob_files <glob>)` and `(glob_files_rec <glob>)` terms to the `files`
+  field of the `install` stanza (#6250, closes #6018, @gridbugs)
+
 3.5.0 (2022-10-19)
 ------------------